Rename detection to analysis and move to licensedb
